// src/shared/api/models/getApiCampaigns200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiCampaigns200CampaignsItem } from './getApiCampaigns200CampaignsItem';

export type GetApiCampaigns200 = {
    campaigns: GetApiCampaigns200CampaignsItem[];
    total: number;
    page: number;
    limit: number;
    totalPages: number;
};