// src/shared/api/models/getApiCampaignsParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiCampaignsCampaignType } from './getApiCampaignsCampaignType';
import type { GetApiCampaignsCampaignStatusItem } from './getApiCampaignsCampaignStatusItem';
import type { GetApiCampaignsSortBy } from './getApiCampaignsSortBy';
import type { GetApiCampaignsSortOrder } from './getApiCampaignsSortOrder';

export type GetApiCampaignsParams = {
    /**
     * @minimum 1
     */
    page?: number;
    /**
     * @minimum 1
     * @maximum 10000
     */
    limit?: number;
    /**
     * Filter by campaign type
     */
    campaign_type?: GetApiCampaignsCampaignType;
    /**
     * Filter by campaign status (multiple selection). Accepts array of statuses: "DRAFT", "SCHEDULED", "TRIGGERED", "COMPLETED", "FAILED", "PAUSED", "CANCELLED"
     */
    campaign_status?: GetApiCampaignsCampaignStatusItem[];
    /**
     * Field name to sort by
     */
    sortBy?: GetApiCampaignsSortBy;
    /**
     * Sort order: ASC or DESC
     */
    sortOrder?: GetApiCampaignsSortOrder;
};